# backend/models/model_resource_list.go
/*
 * KubernetesAccessManager - API
 *
 * This is a backend API server documentation for KubernetesAccessManager  Some useful links: - [Jira](https://samuelus.atlassian.net/jira/software/projects/ZPI/boards/4) - [Confluence](https://samuelus.atlassian.net/wiki/spaces/ZPI/overview)
 *
 * API version: 0.0.2
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package models

// Object that returns selected columns and their data.
type ResourceList struct {
	// List of column names that are included in this response.
	Columns []string `json:"columns,omitempty"`
	// List of resources, each containing the data for the specified columns.
	ResourceList []ResourceListResourceList `json:"resource_list,omitempty"`
}